requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
requests==2.31.0
nltk==3.9.1
python-dotenv==1.0.1